+++ packages/widget/CHANGELOG.md
@@ -1,5 +1,25 @@
 ## 0.0.15
 
+## 3.3.4
+
+### Patch Changes
+
+- d306a00: hotfix noble to penumbra address
+
+## 3.3.3
+
+### Patch Changes
+
+- 667c41a: support noble intermediary chain to penumbra
+
+## 3.3.2
+
+### Patch Changes
+
+- 04fe86d: penumbra wallet address fixes
+- Updated dependencies [04fe86d]
+  - @skip-go/client@0.16.11
+
 ## 3.3.1
 
 ### Patch Changes